Ensure atomic EBF stdout

+++ sw/ecf/embedded_middleware_framework/src/emf_crc/src/emf_crc.c
@@ -107,7 +107,7 @@ uint16_t EMF_crc_16CCITT(const uint8_t* data, uint16_t length)
     {
       if (crc & CRC16_TOPBIT)
       {
-        crc = (uint16_t)((crc << 1) ^ CRC16_CCITT_POLY);
+        crc = (uint16_t)(((uint32_t)crc << 1) ^ CRC16_CCITT_POLY);
       }
       else
       {
